Prod test v1

+++ app/Services/Rating/RatingSkillService.php
@@ -2,6 +2,12 @@
 
 namespace App\Services\Rating;
 
+use App\Exceptions\Rating\CantUpdateValidatedRatingException;
+use App\Exceptions\Rating\DefaultSkillNeedsANameException;
+use App\Exceptions\Rating\SkillAlreadyExistException;
+use App\Exceptions\Rating\SkillsExceedsMarkingException;
+use App\Exceptions\UnauthorizedActionException;
+use App\Models\Rating\Rating;
 use App\Models\Rating\RatingSkill;
 use App\Models\Settings\SkillType;
 
@@ -11,29 +17,49 @@ readonly class RatingSkillService
     {
     }
 
-    public function update($validated, $id): string
+    /**
+     * @throws UnauthorizedActionException
+     * @throws CantUpdateValidatedRatingException
+     */
+    public function update($validated, $id): void
     {
         $ratingSkill = RatingSkill::findOrFail($id);
-        if ($ratingSkill->rating->evaluator_id !== \Auth::id()) return 'revoked';
+        if ((new ValidatorService())->checkForAllValidation(Rating::findOrFail($ratingSkill->rating_id))) throw new CantUpdateValidatedRatingException();
+        if ($ratingSkill->rating->evaluator_id !== \Auth::id()) throw new UnauthorizedActionException();
 //        if(!$this->ratingService->checkMarking($ratingSkill->rating_id)) return "exceed";
         $this->ratingService->lowerMark($ratingSkill->rating_id, $ratingSkill->rating_skill_mark);
         $ratingSkill->update(['rating_skill_mark' => $validated['rating_skill_mark']]);
         $this->ratingService->raiseMark($ratingSkill->rating_id, $ratingSkill->rating_skill_mark);
-        return "ok";
     }
 
-    public function create(mixed $validated): string
+    /**
+     * @throws SkillsExceedsMarkingException
+     * @throws SkillAlreadyExistException
+     * @throws DefaultSkillNeedsANameException
+     * @throws CantUpdateValidatedRatingException
+     */
+    public function create(mixed $validated): void
     {
-        if (!$this->ratingService->checkMarking($validated['rating_id'])) return "exceed";
-//        if (RatingSkill::where('rating_id', '=', $validated['rating_id'])->where('phase_skill_id', '=', $validated['phase_skill_id'])->exists()) return "exist";
-        if (RatingSkill::where('rating_id', '=', $validated['rating_id'])->where('rating_skill_name', '=', $validated['rating_skill_name'])->exists()) return "exist";
+        if ((new ValidatorService())->checkForAllValidation(Rating::findOrFail($validated['rating_id']))) throw new CantUpdateValidatedRatingException();
+        if (!$this->ratingService->checkMarking($validated['rating_id'])) throw new SkillsExceedsMarkingException();
+        if ($validated['phase_skill_id'] !== 1) {
+            $validated['rating_skill_name'] = '';
+            if (RatingSkill::where('rating_id', '=', $validated['rating_id'])->where('phase_skill_id', '=', $validated['phase_skill_id'])->exists()) throw new SkillAlreadyExistException();
+        } else {
+            if (!isset($validated['phase_skill_name'])) throw new DefaultSkillNeedsANameException();
+            if (RatingSkill::where('rating_id', '=', $validated['rating_id'])->where('rating_skill_name', '=', $validated['rating_skill_name'])->exists()) throw new SkillAlreadyExistException();
+
+        }
         RatingSkill::create($validated);
-        return "ok";
     }
 
+    /**
+     * @throws CantUpdateValidatedRatingException
+     */
     public function delete(string $rating_id): void
     {
         $ratingSkill = RatingSkill::with('phaseSkill')->findOrFail($rating_id);
+        if ((new ValidatorService())->checkForAllValidation(Rating::findOrFail($ratingSkill->rating_id))) throw new CantUpdateValidatedRatingException();
         $this->ratingService->lowerMark($ratingSkill->rating_id, $ratingSkill->rating_skill_mark);
         $ratingSkill->delete();
     }
